+++ Code/CAI_2ModConfig.lua
@@ -3,7 +3,7 @@
 -- All rights reserved, duplication and modification prohibited.
 -- You may not copy it, package it, or claim it as your own.
 -- Created Dec 24th, 2018
-
+-- Updated Dec 29th, 2018
 
 local lf_print = false -- Setup debug printing in local file